+++ example.py
@@ -1,53 +1,34 @@
-
 from pecrs import *
 import pyglet
 
 
-class Dude(pyglet.sprite.Sprite):
-   def __init__(self, *args, **kwargs):
-      super().__init__(*args, **kwargs)
-      self.id = None
-      self.speed = 100
-      self.moving = True
-      self.direction = None
-      self.area = None
-
-
-class Objects(Controller):
-   def __init__(self, batch):
-      super().__init__()
-      self.batch = batch
-      self.blue_image = pyglet.resource.image("blue_rect.png")
-      self.red_image = pyglet.resource.image("red_rect.png")
-
-   def make_dude(self, x, y, dx=0, dy=0):
-      body = Dude(self.blue_image, x=x, y=y, batch=self.batch)
-      body.direction = (dx, dy)
-      self.add(body)
-
-   def on_collision(self, body, collisions):
-      body.image = self.red_image
-      
-
-class Game:
+class World(Space):
    def __init__(self):
+      super().__init__()
       self.window = pyglet.window.Window(400, 300)
       self.batch = pyglet.graphics.Batch()
+      
+      self.red_image = pyglet.resource.image("red_rect.png")
+      self.blue_image = pyglet.resource.image("blue_rect.png")
 
-      self.objects = Objects(self.batch)
-      self.objects.make_dude(0, 150, dx=1)
-      self.objects.make_dude(300, 150)
+      spriteA = pyglet.sprite.Sprite(self.blue_image, x=0, y=150, batch=self.batch)
+      spriteB = pyglet.sprite.Sprite(self.blue_image, x=300, y=150, batch=self.batch)
 
+      self.add(spriteA)
+      self.turn(spriteA, (150, 0))
+      self.start_moving(spriteA)
+
+      self.add(spriteB)
+      
       pyglet.clock.schedule_interval(self.run, 1.0/60)
       pyglet.app.run()
 
+   def on_collision(self, shape, collisions):
+      shape.image = self.red_image
+
    def run(self, delta):
-      self.objects.step(delta)
+      self.step(delta)
       self.window.clear()
       self.batch.draw()
 
-game = Game()
-
-
-
-
+world = World()