+++ pecrs/index.py
@@ -12,7 +12,6 @@ class Index:
    def __init__(self):
       self.count = -1 #: The ID to be used for the next new body
       self.free = [] #: A list of free IDs for reuse
-      self.list = {} #: List of objects by thier id
       
    def next(self):
       """
@@ -27,18 +26,6 @@ class Index:
       else:
          return self.free.pop()
 
-   def add(self, item, id=None):
-      if id == None:
-         id = self.next()
-      item.id = id
-      self.list[id] = item
-
-   def get(self, id):
-      return self.list[id]
-
-   def has(self, id):
-      return id in self.list
-
    def delete(self, id):
       """
       :param id: Previously assigned id to be returned to the system
@@ -46,7 +33,6 @@ class Index:
 
       Frees an identifer to be reused.
       """
-      if id in self.list:
-         del self.list[id]
+      if id <= self.count:         
          self.free.append(id)