+++ pecrs/space.py
@@ -1,6 +1,7 @@
 
 from pecrs.spatialhash import SpatialHash
 from pecrs.collider import Collider
+from pecrs.index import Index
 
 class Space:
    """
@@ -14,51 +15,71 @@ class Space:
    def __init__(self, size=256):
       self.grid = SpatialHash(size) #: SpatialHash, spatial partitioning system. Think about combining these into one class
       self.collider = Collider() #: Collider, collision detection system
+      self.index = Index()
+      self.ids = {}
+      self.shapes = {}
+      self.areas = {}
+      self.moving = {}
+      self.directions = {}
+      self.actives = []
+      self.groups = {}
 
-   def add(self, shape, update_area=True):
+   def add(self, shape, id=-1, active=True, group=None):
       """
       :param shape: Shape to add to the space
       :type shape: Shape
       
       Adds a shape to the space.
-      Bodies are added to their collision area as well as the 8 nearest neighboring cells to handle overlap and fast moving obejcts.
       """
-      if update_area:
-         shape.area = self.grid.scale(shape.position[0], shape.position[1])
+      if id < 0:
+         id = self.index.next()
+      self.ids[shape] = id
+      self.shapes[id] = shape
+      self.moving[shape] = False
+      area = self.grid.scale(shape.position[0], shape.position[1])   
+      self.areas[shape] = area
+      self.add_to_grid(shape, area)
+      
+      if active:
+         self.actives.append(shape)
+         self.directions[shape] = (0, 0)
+      
+      if group:
+         if group in self.groups:
+            self.groups[group].append(shape)
+         else:
+            self.groups[group] = [shape]
 
-      x_area = shape.area[0]
-      y_area = shape.area[1]
-
-      self.grid.add(shape, (x_area-1, y_area-1))
-      self.grid.add(shape, (x_area, y_area-1))
-      self.grid.add(shape, (x_area+1, y_area-1))
-      self.grid.add(shape, (x_area-1, y_area))
-      self.grid.add(shape, (x_area, y_area))
-      self.grid.add(shape, (x_area+1, y_area))
-      self.grid.add(shape, (x_area-1, y_area+1))
-      self.grid.add(shape, (x_area, y_area+1))
-      self.grid.add(shape, (x_area+1, y_area+1))
-
-   def delete(self, shape):
+   def delete(self, shape, group=None):
       """
       :param shape: Shape to removed from the space
       :type shape: Shape
 
       Removes a shape from the space.
       """
-      x_area = shape.area[0]
-      y_area = shape.area[1]
+      if shape in self.ids:
+         id = self.ids[shape]
+         del self.ids[shape]
+         del self.shapes[id]
+         del self.moving[shape]
+         self.delete_from_grid(shape, self.areas[shape])
+         del self.areas[shape]
+         if shape in self.actives: #TODO I think self.actives, self.directions might be work better with groups if given?
+            self.actives.remove(shape)
+         if shape in self.directions:
+            del self.directions[shape]
+         if group:
+            self.groups[group].remove(shape)
 
-      self.grid.delete(shape, (x_area-1, y_area-1))
-      self.grid.delete(shape, (x_area, y_area-1))
-      self.grid.delete(shape, (x_area+1, y_area-1))
-      self.grid.delete(shape, (x_area-1, y_area))
-      self.grid.delete(shape, (x_area, y_area))
-      self.grid.delete(shape, (x_area+1, y_area))
-      self.grid.delete(shape, (x_area-1, y_area+1))
-      self.grid.delete(shape, (x_area, y_area+1))
-      self.grid.delete(shape, (x_area+1, y_area+1))
+   def delete_id(self, id, group=None):
+      shape = self.ids[shape]
+      self.delete(shape, group)
 
+   def delete_group(self, group):
+      for shape in self.groups[group]:
+         self.delete(shape)
+      del self.groups[group]
+      
    def has(self, shape):
       """
       :param shape: Shape to check for
@@ -69,10 +90,10 @@ class Space:
       Checks if a space has a shape in that shape's area
       Note that the shape's must be set correctly or errors can occur
       """
-      try:
-         return self.grid.has(shape, shape.area)
-      except AttributeError:
-         return False
+      return shape in self.ids
+
+   def has_id(self, id):
+      return id in self.shapes
 
    def collisions_at(self, x, y, width=1, height=1):
       """
@@ -89,11 +110,10 @@ class Space:
 
       Gets a list of all shapes colliding at x, y
       """
-      area = self.grid.scale(x, y)
       collisions = []
-      for shape in self.grid.get(area):
-         if self.collider.rect_rect(width, height, x, y, shape.width, shape.height, shape.position[0], shape.position[1]):
-            collisions.append(shape)
+      for other in self.grid.get(self.grid.scale(x, y)):
+         if self.collider.rect_rect(width, height, x, y, other.width, other.height, other.position[0], other.position[1]):
+            collisions.append(other)
       return collisions
 
    def collisions_with(self, shape):
@@ -106,12 +126,26 @@ class Space:
       Get a list of all shapes colliding with shape
       """
       collisions = []
-      for other in self.grid.get(shape.area):
-         if shape != other:
-            if self.collider.check_rects(shape, other):
+      id = self.ids[shape]
+      for other in self.grid.get(self.areas[shape]):
+         if id != self.ids[other]:
+            if self.collider.rect_rect(shape.width, shape.height, shape.position[0], shape.position[1], other.width, other.height, other.position[0], other.position[1]):
                collisions.append(other)
       return collisions
-      
+
+   def collisions_with_group(self, group):
+      collisions = []
+      group = self.groups[group]
+      for shape in group:
+         id = self.ids[shape]
+         for other in self.grid.get(self.areas[shape]):
+            other_id = self.ids[other]
+            if id != self.ids[other]:
+               if other not in group:
+                  if self.collider.rect_rect(shape.width, shape.height, shape.position[0], shape.position[1], other.width, other.height, other.position[0], other.position[1]):
+                     collisions.append(other)
+      return collisions
+
    def check_at(self, x, y, width=1, height=1):
       """
       :param x: Posistion to search for on the x-axis
@@ -127,9 +161,8 @@ class Space:
 
       Check to see if there is a collision at x, y in space.
       """
-      area = self.grid.scale(x, y)
-      for shape in self.grid.get(area):
-         if self.collider.rect_rect(width, height, x, y, shape.width, shape.height, shape.position[0], shape.position[1]):
+      for other in self.grid.get(self.grid.scale(x, y)):
+         if self.collider.rect_rect(width, height, x, y, other.width, other.height, other.position[0], other.position[1]):
             return True
       return False
       
@@ -142,11 +175,24 @@ class Space:
 
       Check to see if something is colliding with shape in the space.
       """
-      for other in self.grid.get(shape.area):
-         if shape != other:
-            if self.collider.check_rects(shape, other):
+      id = self.ids[shape]
+      for other in self.grid.get(self.areas[shape]):
+         if id != self.ids[other]:
+            if self.collider.rect_rect(shape.width, shape.height, shape.position[0], shape.position[1], other.width, other.height, other.position[0], other.position[1]):
                return True
       return False
+
+   def check_group(self, group):
+      group = self.groups[group]
+      for shape in group:
+         id = self.ids[shape]
+         for other in self.grid.get(self.areas[shape]):
+            if id != self.ids[other]:
+               if other not in group:
+                  if self.collider.rect_rect(shape.width, shape.height, shape.position[0], shape.position[1], other.width, other.height, other.position[0], other.position[1]):
+                     return True
+      return False
+
 
    def check_two(self, shape, other):
       """
@@ -173,16 +219,33 @@ class Space:
       Moves a shape from one collision area to another
       """
       area = self.grid.scale(shape.position[0], shape.position[1])
-      if shape.area != area:
-         self.delete(shape)
-         start = shape.area
-         shape.area = area
-         self.add(shape, False)
+      start = self.areas[shape]
+      if start != area:
+         self.move_area(shape, area)
          return start
       else:
          return None
+
+   def move_area(self, shape, area):
+      #TODO this needs to be optimized better but I'm not doing it right now
+      start = self.areas[shape]
+      self.delete_from_grid(shape, area)
+      self.add_to_grid(shape, area)
+      self.areas[shape] = area
+
+   def delete_from_grid(self, shape, area):
+      self.grid.delete(shape, area)
+      self.grid.delete(shape, (area[0]+1, area[1]))
+      self.grid.delete(shape, (area[0], area[1]+1))
+      self.grid.delete(shape, (area[0]+1, area[1]+1))
+
+   def add_to_grid(self, shape, area):
+      self.grid.add(shape, area)
+      self.grid.add(shape, (area[0]+1, area[1]))
+      self.grid.add(shape, (area[0], area[1]+1))
+      self.grid.add(shape, (area[0]+1, area[1]+1))
       
-   def move(self, shape, x, y, distance):
+   def move(self, shape, distance):
       """
       :param shape: Shape to move
       :param x: Direction to move in on horizontal plane
@@ -195,12 +258,18 @@ class Space:
       :return: Distance moved
       :rtype: tuple(int, int)
       
-      Advances a shape in the direction x, y for distance units.
+      Advances a shape in its direction for distance units.
       """
-      xstep = int(x * distance)
-      ystep = int(y * distance)
+      direction = self.directions[shape]
+      xstep = int(direction[0] * distance)
+      ystep = int(direction[1] * distance)
       shape.position = (shape.position[0]+xstep, shape.position[1]+ystep)
+      self.update_area(shape)
       return (xstep, ystep)
+
+   def move_group(self, group, distance):
+      for shape in self.groups[group]:
+         self.move(shape, distance)
 
    def push(self, shape, x, y):
       """
@@ -214,6 +283,11 @@ class Space:
       Pushes a shape in the direction of x, y in space
       """
       shape.position = (shape.position[0]+x, shape.position[1]+y)
+      self.update_area(shape)
+
+   def push_group(self, group, x, y):
+      for shape in self.groups[group]:
+         self.push(shape, x, y)
 
    def place(self, shape, x, y):
       """
@@ -225,4 +299,88 @@ class Space:
       Directly places a shape at x, y in space
       """
       shape.position = (x, y)
+      self.update_area(shape)
+
+   def place_group(self, group, x, y):
+      for shape in self.groups[group]:
+         self.place(shape, x, y)
+
+   def start_moving(self, shape):
+      self.moving[shape] = True
+
+   def start_group(self, group):
+      for shape in self.groups[group]:
+         self.moving[shape] = True
+         
+   def stop_moving(self, shape):
+      self.moving[shape] = False
+
+   def stop_group(self, group):
+      for shape in self.groups[group]:
+         self.moving[shape] = False
+         
+   def turn(self, shape, direction):
+      self.directions[shape] = direction
+
+   def turn_group(self, group, direction):
+      for shape in self.groups[group]:
+         self.directions[shape] = direction
+
+
+   def step(self, delta):
+      """
+      :param delta: Units of time to advance the simulation
+      :type delta: float
+      Process the system for delta steps.
+      Triggers the on_step_start(delta), on_step(body, delta), on_collision(body, collisions), and on_step_end(delta) callbacks
+      """
+      self.on_step_start(delta)
+      for shape in self.actives:
+         self.on_step(shape, delta)
+         if self.moving[shape]:
+            self.move(shape, delta)
+
+      #TODO This is very unoptimized
+      for shape in self.actives:
+         collisions = self.collisions_with(shape)
+         if collisions:
+            self.on_collision(shape, collisions)
+      self.on_step_end(delta)
+
+   def on_step_start(self, delta):
+      """
+      :param delta: Units of time to advance the simulation
+      :type delta: float
+      Callback before the simulation takes a step. Override this when extending your Controller
+      """
+      return
+
+
+   def on_step(self, shape, delta):
+      """
+      :param body: Body being processed by the step
+      :param delta: Units of time to advance the simulation
+      :type body: Body
+      :type delta: float
+      Callback when the simulation takes a step over a body. Override this when extending your Controller
+      """
+      return
+
+   def on_collision(self, shape, collisions):
+      """
+      :param body: Body involved in a collision
+      :param collision: List of bodies colliding with body
+      :type body: Body
+      :type other: List(Body)
+      Callback when bodies collide with another body. Override this when extending your Controller
+      """
+      return
+
+   def on_step_end(self, delta):
+      """
+      :param delta: Units of time to advance the simulation
+      :type delta: float
+      Callback after the simulation takes a step. Override this when extending your Controller
+      """
+      return