+++ test.py
@@ -1,6 +1,11 @@
-import pecrs
-from pecrs import *
-
+from pecrs.index import Index
+from pecrs.spatialhash import SpatialHash
+from pecrs.space import Space
+from pecrs.shape import Rect
+from pecrs.shape import Circle
+from pecrs.collider import Collider
+from pecrs.clock import Clock
+from pecrs.clock import SyncClock
 
 def test_index():
   passing = True
@@ -19,42 +24,45 @@ def test_index():
     print("Fail")
     passing = False
 
-  print(f"Index test: add(Rect(), {a})")
-  index.add(Rect(0, 0, 32, 32), a)
+  c = index.next()
+  print(f"Index test: next() expecting 2, actual {c}")
+  if c != 2:
+    print("Fail")
+    passing = False
 
-  print(f"Index test: add(Rect(), {b})")
-  index.add(Rect(0, 0, 32, 32), b)
-
-  print(f"Index test: add(Rect())")
-  index.add(Rect(0, 0, 32, 32))
+  d = index.next()
+  print(f"Index test: next() expecting 3, actual {d}")
+  if d != 3:
+    print("Fail")
+    passing = False
 
   index.delete(a)
   print(f"Index test: delete({a})")
 
-  d = index.next()
-  print(f"Index test: next() expecting 0, actual {d}")
-  if d != 0:
+  e = index.next()
+  print(f"Index test: next() expecting 0, actual {e}")
+  if e != 0:
     print("Fail")
     passing = False
 
   index.delete(b)
   print(f"Index test: delete({b})")
 
-  e = index.next()
-  print(f"Index test: next() expecting 1, actual {e}")
-  if e != 1:
-    print("Fail")
-    passing = False
-
   f = index.next()
-  print(f"Index test: next() expecting 3, actual {f}")
-  if f != 3:
+  print(f"Index test: next() expecting 1, actual {f}")
+  if f != 1:
     print("Fail")
     passing = False
 
   g = index.next()
   print(f"Index test: next() expecting 4, actual {g}")
   if g != 4:
+    print("Fail")
+    passing = False
+
+  h = index.next()
+  print(f"Index test: next() expecting 5, actual {h}")
+  if h != 5:
     print("Fail")
     passing = False
 
@@ -344,14 +352,13 @@ def test_space():
     passing = False
 
   space.add(rectA)
-  print("Space test: add(rectA)")
-  rectB.area = (0, 0)
-  space.add(rectB, False)
-  print("Space test: add(rectB, False)")
+  print(f"Space test: add(rectA) id {space.ids[rectA]}: {rectA}")
+  space.add(rectB)
+  print(f"Space test: add(rectB) id {space.ids[rectB]}: {rectB}")
   space.add(rectC)
-  print("Space test: add(rectC)")
+  print(f"Space test: add(rectC) id {space.ids[rectC]}: {rectC}")
   space.add(rectD)
-  print("Space test: add(rectD)")
+  print(f"Space test: add(rectD) id {space.ids[rectD]}: {rectD}")
 
   value = space.has(rectA)
   print(f"Space test: has(rectA)) expecting True, actual {value}")
@@ -359,14 +366,8 @@ def test_space():
     print("Fail")
     passing = False
 
-  value = space.has(rectA)
-  print(f"Space test: has(rectB)) expecting True, actual {value}")
-  if value != True:
-    print("Fail")
-    passing = False
-
   value = space.collisions_at(0, 0, 32, 32)
-  print(f"Space test: collisions_at(0, 0, 32, 32) expecting [...], actual [...]")
+  print(f"Space test: collisions_at(0, 0, 32, 32) expecting [{rectA}, {rectB}]")
   for v in value:
     if v == rectA or v == rectB:
       pass
@@ -375,33 +376,33 @@ def test_space():
       passing = False
 
   value = space.collisions_at(0, 64)
-  print(f"Space test: colliding_at(0, 64) expecting [...], actual [...]")
+  print(f"Space test: colliding_at(0, 64) expecting [{rectC}]")
   for v in value:
     if v != rectC:
-      print("Fail")
+      print(f"Fail {v}")
       passing = False
 
   value = space.collisions_with(rectA)
-  print(f"Space test: collisions_with(rectA) expecting [...], actual [...]")
+  print(f"Space test: collisions_with(rectA) expecting [{rectB}]")
   for v in value:
     if v != rectB:
-      print("Fail")
+      print(f"Fail {v}")
       passing = False
 
   value = space.collisions_with(rectD)
-  print(f"Space test: collisions_with(rectD) expecting [...], actual [...]")
+  print(f"Space test: collisions_with(rectD) expecting []")
   if value != []:
-    print("Fail")
+    print(f"Fail {value}")
     passing = False
 
   value = space.check_two(rectA, rectB)
-  print(f"Space test: check_two(rectA, rectB) expecting True, actual {value}")
+  print(f"Space test: check_two(rectA, rectB) expecting True")
   if value != True:
     print("Fail")
     passing = False
 
   value = space.check_two(rectC, rectD)
-  print(f"Space test: check_bodies(rectC, rectD) expecting False, actual {value}")
+  print(f"Space test: check_two(rectC, rectD) expecting False")
   if value != False:
     print("Fail")
     passing = False
@@ -422,124 +423,103 @@ def test_space():
   print("Space test: place(rectD, 0, 0)")
 
   value = space.collisions_with(rectA)
-  print(f"Space test: collisions_with(rectA) expecting [...], actual [...]")
+  print(f"Space test: collisions_with(rectA) expecting [{rectB}, {rectD}]")
   for v in value:
     if v == rectB or v == rectD:
       pass
     else:
-      print("Fail")
+      print(f"Fail {v}")
       passing = False
-
+  
+  area = space.areas[rectA]
   space.delete(rectA)
   print("Space test: delete(rectA)")
+  value = space.grid.get(area)
+  for v in value:
+    if v == rectA:
+      print("Fail")
 
   value = space.collisions_with(rectB)
-  print(f"Space test: collisions_with(rectB) expecting [...], actual [...]")
+  print(f"Space test: collisions_with(rectB) expecting [{rectD}]")
   for v in value:
-    if v != rectD:
-      print("Fail")
+    if v == rectD:
+      pass
+    else:
+      print(f"Fail {v}")
       passing = False
 
-  if passing:
-    print("Space test: Complete, all tests passed")
-  else:
-    print("Space test: Incomplete, some tests failed")
-  return passing
+  print("Space test group testing")
+  space = Space()
+  headA = Rect(0, 0, 32, 32)
+  bodyA = Rect(0, 30, 32, 32)
+  headB = Rect(64, 0, 32, 32)
+  bodyB = Rect(64, 32, 32, 32)
+  space.add(headA, group="a")
+  space.add(bodyA, group="a")
+  space.start_group("a")
+  space.add(headB, group="b")
+  space.add(bodyB, group="b")
 
-
-def test_controller():
-  passing = True
-
-  #TODO test controller.remake_space()
-
-  class Entity(Body):
-    def __init__(self, id, x, y, width, height):
-      super().__init__(id, x, y, width, height)
-
-  class Objects(Controller):
-    def __init__(self):
-      super().__init__()
-
-    def on_add(self, body):
-      print(f"on_add callback {body.id}")
-
-    def on_make(self, body):
-      body.speed = 100
-      body.moving = True
-      print(f"on_make callback {body.id}")
-
-    def on_delete(self, body):
-      print(f"on_delete callback {body.id}")
-
-    def on_area(self, body, start):
-      print(f"on_area callback {body.id} from {start} to {body.area}")
-
-    def on_move(self, body, distance):
-      print(f"on_move callback {body.id} moved {distance} units")
-
-    def on_place(self, body, start):
-      print(f"on_place callback {body.id} moved from {start}")
-
-    def on_motion(self, body):
-      print(f"body {body.id} at {str(body.position)}")
-
-    def on_turn(self, body):
-      print(f"on_turn callback {body.id} to {str(body.direction)}")
-
-    def on_start(self, body):
-      print(f"on_start callback {body.id}")
-
-    def on_stop(self, body):
-      print(f"on_stop callback {body.id}")
-
-    def on_step_start(self, delta):
-      print("on_step_start callback")
-
-    def on_step(self, body, delta):
-      print(f"on_step callback {body.id} {delta}")
-
-    def on_step_end(self, delta):
-      print("on_step_end callback")
-
-  objects = Objects()
-
-  bodyA = objects.make(Entity, 0, 0, 32, 32)
-  print(f"Controller test: make(Entity, 0, 0, 32, 32) expecting id 0, actual {bodyA.id}")
-  if bodyA.id != 0:
+  value = space.collisions_with_group("a")
+  print(f"Space test: collisions_with_group(\"a\") expecting [], actual {value}")
+  if value != []:
     print("Fail")
     passing = False
+  
+  space.move_group("a", 1)
+  print(f"Space test: group a position {headA.position}, {bodyA.position}")
+  print(f"Space test: group b position {headB.position}, {bodyB.position}")
 
-  bodyB = objects.make(Entity, 0, 0, 32, 32, dx=1)
-  print(f"Controller test: make(Entity, 0, 0, 32, 32, dx=1) expecting id 1, actual {bodyB.id}")
-  if bodyB.id != 1:
-    print("Fail")
+  space.place_group("b", 1, 1)
+  print(f"Space test: group a position {headA.position}, {bodyA.position}")
+  print(f"Space test: group b position {headB.position}, {bodyB.position}")
+
+  value = space.collisions_with_group("a")
+  print(f"Space test: collisions_with_group(\"a\") expecting [...], actual {value}")
+  for v in value:
+    if v == headB or v == bodyB:
+      pass
+    else:
+      print(f"Fail {v}")
+      passing = False
+  
+  space.turn_group("a", (1, 0))
+  space.turn(bodyB, (-1, 0))
+  space.turn(headB, (-1, 0))
+
+  space.step(1)
+  print(f"Space test: group a position {headA.position}, {bodyA.position}")
+  print(f"Space test: group b position {headB.position}, {bodyB.position}")
+
+  value = space.collisions_with_group("b")
+  print(f"Space test: collisions_with_group(\"b\") expecting [...], actual {value}")
+  for v in value:
+    if v == headA or v == bodyA:
+      pass
+    else:
+      print(f"Fail {v}")
+      passing = False
+
+  space.step(10)
+  print(f"Space test: group a position {headA.position}, {bodyA.position}")
+  print(f"Space test: group b position {headB.position}, {bodyB.position}")
+  
+  value = space.collisions_with_group("a")
+  print(f"Space test: collisions_with_group(\"a\") expecting [...], actual {value}")
+  for v in value:
+    if v == headB or v == bodyB:
+      pass
+    else:
+      print(f"Fail {v}")
+      passing = False
+
+  space.delete_group("b") 
+
+  value = space.collisions_with_group("a")
+  print(f"Space test: collisions_with_group(\"a\") expecting [...], actual {value}")
+  for v in value:
+    print(f"Fail {v}")
     passing = False
-
-  bodyC = objects.make(Entity, 0, 0, 32, 32, dy=1)
-  print(f"Controller test: make(Entity, 0, 0, 32, 32 dy=1) expecting id 2, actual {bodyC.id}")
-  if bodyC.id != 2:
-    print("Fail")
-    passing = False
-
-  bodyD = objects.make(Entity, 0, 0, 32, 32)
-  print(f"Controller test: make(Entity, 0, 0, 32, 32) expecting id 3, actual {bodyD.id}")
-  if bodyD.id != 3:
-    print("Fail")
-    passing = False
-
-  objects.move(bodyA, 1)
-  objects.move(bodyB, 1)
-  objects.move_to(bodyD, 0, 1, 1)
-  objects.place(bodyC, 10, 10)
-
-  objects.turn(bodyD, -1, 0)
-
-  objects.step(1)
-
-  objects.delete(bodyA)
-  objects.delete_id(bodyD.id)
-
-  objects.step(1)
 
   if passing:
     print("Space test: Complete, all tests passed")
@@ -561,13 +541,10 @@ def test():
   if not test_space():
     passing = False
 
-  if not test_controller():
-    passing = False
-
   if passing:
-    print(f"pecrs version {pecrs.version} test: Complete, all tests passed")
+    print("pecrs version {pecrs.version} test: Complete, all tests passed")
   else:
-    print(f"pecrs version {pecrs.version} test: Incomplete, some tests failed")
+    print("pecrs version {pecrs.version} test: Incomplete, some tests failed")
 
 test()